requirements.txt
transformers
Pillow
torch
torchvision
matplotlib
numpy